[BE/POST] PostException 처리 및 코드 전체 리팩터링

+++ Backend/src/main/java/com/ssafy/backend/post/service/PostServiceImpl.java
@@ -5,6 +5,10 @@ import com.ssafy.backend.cafe.domain.dto.NearByCafeResultDto;
 import com.ssafy.backend.cafe.domain.entity.Cafe;
 import com.ssafy.backend.cafe.repository.CafeRepository;
 import com.ssafy.backend.cafe.service.CafeServiceImpl;
+import com.ssafy.backend.common.exception.member.MemberException;
+import com.ssafy.backend.common.exception.member.MemberExceptionType;
+import com.ssafy.backend.common.exception.post.PostException;
+import com.ssafy.backend.common.exception.post.PostExceptionType;
 import com.ssafy.backend.member.domain.entity.Member;
 import com.ssafy.backend.member.domain.entity.MemberCafeTier;
 import com.ssafy.backend.member.repository.MemberCafeTierRepository;
@@ -26,6 +30,8 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.transaction.Transactional;
 import java.util.*;
 
+import static com.ssafy.backend.common.exception.post.PostExceptionType.NO_CONTENT_POST_FORM;
+
 @RequiredArgsConstructor // 얘도 커스텀?
 @Service
 @Transactional
@@ -34,21 +40,20 @@ public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
     private final PostLikeRepository postLikeRepository;
     private final MemberRepository memberRepository;
-    private final CommentRepository commentRepository;
-    // Util
-    private final PostUtil postUtil;
-    private final CafeServiceImpl cafeService;
-    private final MemberServiceImpl memberService;
     private final CafeAuthRepository cafeAuthRepository;
     private final CafeRepository cafeRepository;
     private final MemberCafeTierRepository memberCafeTierRepository;
     private final PostCafeRepository postCafeRepository;
+    // Util
+    private final PostUtil postUtil;
+    private final CafeServiceImpl cafeService;
+    private final MemberServiceImpl memberService;
 
     /**
      * 1. 글 등록 [ 테스트 완료 ]
      **/
     @Override
-    public Long writePost(MultipartFile[] files, PostWriteFormRequestDto requestDto) throws Exception {
+    public Long writePost(MultipartFile[] files, PostWriteFormRequestDto requestDto) {
 
         //1. 유저 확인
         CheckedResponseDto checked = postUtil.checkMember();
@@ -56,21 +61,28 @@ public class PostServiceImpl implements PostService {
         Double latitude = requestDto.getLatitude();
         Double longitude = requestDto.getLongitude();
         Double dist = requestDto.getDist();
-
-        // 1-2. 글 저장하기
-
-        Optional<Member> optionalMember = memberRepository.findById(memberId);
-        Member member = optionalMember.orElseThrow();
         String content = requestDto.getContent();
-        System.out.println("content : " + content);
+        if (files != null || content != null) {
+
+        } else {
+            throw new PostException(NO_CONTENT_POST_FORM); // 이미지나 글 둘중 하나라도 없으면 에러
+        }
+
+        // 1-2. 멤버 불러오기
+        Optional<Member> optionalMember = memberRepository.findById(memberId);
+        if (optionalMember.isEmpty() || optionalMember == null)
+            throw new MemberException(MemberExceptionType.NOT_FOUND_MEMBER);
+        Member member = optionalMember.get();
 
         // 1-3. 이미지 업로드 Build
-        Post post;
-        post = Post.postWriteBuilder()
+        Post post = Post.postWriteBuilder()
                 .member(member)
-                .content(content)
                 .postType(requestDto.getType())
                 .build();
+
+        if (content != null) {
+            post.updateContent(content);
+        }
 
         if (files != null) {
             List<PostImage> postImages = postUtil.imageUpload(files);
@@ -81,6 +93,7 @@ public class PostServiceImpl implements PostService {
 
         // 1-4. 유저 인증여부 확인
         Optional<CafeAuth> cafeAuth = cafeAuthRepository.findById(checked.getNickname());
+
         PostCafe postCafe;
         if (cafeAuth.isEmpty() || cafeAuth == null) { // 카페 이름이 없으면 - 인증되지 않은 유저
             if (requestDto.getType() == PostType.qna || requestDto.getType() == PostType.lost) { // 카테고리가 둘중 하나면 넘어가기
@@ -91,26 +104,22 @@ public class PostServiceImpl implements PostService {
                             .post(post)
                             .cafe(cafeRepository.findById(resultDto.getId().longValue()).get())
                             .build();
-                    post.addPostCafe(postCafe);
+                    postCafeRepository.save(postCafe); // Post 가 너무 뚱뚱해지는걸 막기위해
                 }
 
-
             } else { // 그렇지 않다면 false
-                return -1L;
+                throw new PostException(PostExceptionType.NOT_ALLOWED_POST_TYPE);
             }
         } else { // 카페이름이 있으면 - 인증된 유저
             Cafe cafe = cafeRepository.findById(cafeAuth.get().getCafeId()).get(); // 카페 닉네임을 확인한다.
-            String cafeNickname = cafe.getName();
-
+            post.updateAuthorized();
             // 1-3. 카페위치 저장하기
-
             postCafe = PostCafe.PostCafeBuilder()
                     .post(post)
                     .cafe(cafe)
                     .build();
-            post.addPostCafe(postCafe);
+            post.addPostCafe(postCafe); // 인증된 카페의 경우, post 에 postCafe 저장해주기
         }
-
         postRepository.save(post);
         return post.getId();
     }
@@ -121,13 +130,19 @@ public class PostServiceImpl implements PostService {
 
     // 3-1 게시글 하나 삭제
     @Override
-    public boolean deletePost(Long postId) throws Exception {
+    public void deletePost(Long postId) {
+
         Optional<Post> postOptional = postRepository.findById(postId);
-        if(postOptional.isEmpty() || postOptional == null) {
-            return false;
+        if (postOptional.isEmpty() || postOptional == null) {
+            throw new PostException(PostExceptionType.BAD_POST_ID);
+        }
+        Post post = postOptional.get();
+        long memberId = postUtil.checkMember().getMemberId();
+        long writerId = post.getMember().getId();
+        if (memberId != writerId) {
+            throw new PostException(PostExceptionType.USER_IS_NOT_WRITER);
         }
         postRepository.deleteById(postId);
-        return true;
     }
 
     // 3-2 회원탈퇴 시 게시글 모두 삭제
@@ -136,12 +151,11 @@ public class PostServiceImpl implements PostService {
      * 4. 글 전체 조회 (10개) [확인완료]
      **/
     @Override
-    public List<PostPagingResponseDto> feedPosts(PostPagingRequestDto requestDto, Pageable pageable) throws Exception {
+    public List<PostPagingResponseDto> feedPosts(PostPagingRequestDto requestDto, Pageable pageable) {
         String cafeName;
+        Slice<Post> postSlice;
         // 1. 유저 기본사항을 체크한다. OK
         CheckedResponseDto checked = postUtil.checkMember();
-        long memberId = checked.getMemberId(); // 멤버 아이디를 확인한다.
-
 
         // 2. request Dto 값들 체크 - OK
         Long postId = requestDto.getPostId();
@@ -149,28 +163,27 @@ public class PostServiceImpl implements PostService {
         Double latitude = requestDto.getLatitude();
         Double longitude = requestDto.getLongitude();
         Double dist = requestDto.getDist();
-        System.out.println(latitude + " " + longitude + " " + dist);
 
-        // 3. 주변 카페들 정보 알아오기 - 주변 카페에 해당되는 글들만 된다. - Map 객체가 너무 많이 만들어진다. 리팩터링 필요
+        // 3. 주변 카페들 정보 알아오기 - 주변 카페에 해당되는 글들만 된다.
         ClientPosInfoDto clientPosInfoDto = new ClientPosInfoDto(latitude, longitude, dist);
         List<NearByCafeResultDto> nearByCafeResultDtos = cafeService.getNearByCafeLocations(clientPosInfoDto);
 
         if (nearByCafeResultDtos.isEmpty() || nearByCafeResultDtos == null) {
-            System.out.println("주변 카페가 없음?");
+            throw new PostException(PostExceptionType.NO_CAFE_AROUND);
         } else {
-            System.out.println(" 총몇개? : " + nearByCafeResultDtos.size() + "개       " + nearByCafeResultDtos);
+            System.out.println(" 글 불러오기 - 주변 카페 개수 : " + nearByCafeResultDtos.size() + "개       " + nearByCafeResultDtos);
         }
-        // cafe id 와 name 만 전달해줄거임
+        // cafe id 만 전달해줄거임
         List<Long> cafeIdList = new ArrayList<>();
         for (NearByCafeResultDto dto : nearByCafeResultDtos) {
             cafeIdList.add(dto.getId().longValue());
         }
-        Slice<Post> postSlice;
+
         if (types.contains(PostType.hot)) { // 핫 게시물을 포함하고 있을 때
 
             if (postId == -1L) {
                 // 처음 요청할때 (refresh)
-                System.out.println("hot 새로고침");
+                System.out.println("hot 첫번째요청");
                 postSlice = postRepository.findHotPost(cafeIdList, pageable);
             } else {
                 // 두번째 이상으로 요청할 때 (마지막 글의 pk 를 기준으로 함)
@@ -183,7 +196,7 @@ public class PostServiceImpl implements PostService {
 
             if (postId == -1L) {
                 // 처음 요청할때 (refresh)
-                System.out.println("새로고침");
+                System.out.println("피드 첫번째 요청");
                 postSlice = postRepository.findAllByPostTypeInAndPostCafeList_CafeIdIn(types, cafeIdList, pageable);
             } else {
                 // 두번째 이상으로 요청할 때 (마지막 글의 pk 를 기준으로 함)
@@ -196,14 +209,12 @@ public class PostServiceImpl implements PostService {
 //         post를 slice 형태로 갖고오기
 
         if (postSlice.isEmpty() || postSlice == null) { // 불러올 게시물이 있을때
-            System.out.println("널이래용");
-            return null;
+            throw new PostException(PostExceptionType.NO_POST_FEED);
         }
 
         // 5. 리턴값 채워넣기
         List<PostPagingResponseDto> postResponseDtoList = new ArrayList<>();
         for (Post slice : postSlice) {
-
             Optional<Post> optionalPost = postRepository.findById(slice.getId());
             Post post = optionalPost.get();
             List<PostImage> postImages = post.getPostImageList();
@@ -215,8 +226,9 @@ public class PostServiceImpl implements PostService {
             int postLikeCount = post.getPostLikeList().size();
             int commentCount = post.getCommentList().size();
 
+            System.out.println("인증된카페? " + post.isCafeAuthorized());
             PostPagingResponseDto postPagingResponseDto = PostPagingResponseDto.builder()
-                    .postType("미인증 유저의 글")
+                    .isCafeAuthorized(post.isCafeAuthorized())
                     .postId(slice.getId())
                     .imgUrlPath(imgUrlPath)
                     .createdAt(post.getCreatedAt())
@@ -225,26 +237,22 @@ public class PostServiceImpl implements PostService {
                     .writerNickname(post.getMember().getNickname())
                     .postLikeCount(postLikeCount)
                     .build();
-            System.out.println("미인증 유저 글 build 완료");
-                System.out.println("인증된카페? " + post.isCafeAuthorized());
 
             if (post.isCafeAuthorized()) {
+                if (post.getPostCafeList() == null || post.getPostCafeList().isEmpty()) {
+                    throw new PostException(PostExceptionType.NO_POST_CAFE);
+                }
                 // 당연히 없지
-                Optional<PostCafe> postCafeOptional = postCafeRepository.findByPostId(post.getId());
-                PostCafe postCafe = postCafeOptional.get();
+                PostCafe postCafe = post.getPostCafeList().get(0); // 인증된 유저의 글쓰기 경우, postCafe 객체를 하나만 가지고있음
                 Cafe cafe = postCafe.getCafe();
                 Long cafeId = cafe.getId();
-                System.out.println("Cafe 객체는 있음");
                 String brandType = cafe.getBrandType();
                 cafeName = cafe.getName();
-                System.out.println("여기까지도 옴");
                 Long exp = memberCafeTierRepository.findByMemberIdAndCafeId(post.getMember().getId(), cafeId).get().getExp();
-                postPagingResponseDto.updateDto(cafeName,exp,brandType,"인증 유저의 글");
-                System.out.println("여긴오냐?");
+                postPagingResponseDto.updateDto(cafeName, exp, brandType);
             }
             postResponseDtoList.add(postPagingResponseDto);
         }
-
         return postResponseDtoList;
     }
 
@@ -252,18 +260,16 @@ public class PostServiceImpl implements PostService {
      * 4. 상세페이지 조회
      **/
     @Override
-    public PostDetailResponseDto findOnePost(Long postId) throws Exception {
+    public PostDetailResponseDto findOnePost(Long postId) {
         // 1. 로그인된 유저의 정보를 확인한다.
         CheckedResponseDto checked = postUtil.checkMember();
         String nickname = checked.getNickname();
 
         Optional<Post> postOptional = postRepository.findById(postId);
-        Post post;
         if (postOptional == null || postOptional.isEmpty()) {
-            // 잘못된 postId 형식
-            return null;
+            throw new PostException(PostExceptionType.BAD_POST_ID);
         }
-        post = postOptional.get();
+        Post post = postOptional.get();
 
         List<String> imgUrlPaths = new ArrayList<>();
         for (PostImage postImage : post.getPostImageList()) {
@@ -282,23 +288,20 @@ public class PostServiceImpl implements PostService {
                 .commentCounts(post.getCommentList().size())
                 .build();
 
-        System.out.println("postDetail : 미인증 빌드 완료");
-
-        if(detailResponseDto.isCafeAuthorized()) {
+        if (detailResponseDto.isCafeAuthorized()) {
             PostCafe postCafe = postCafeRepository.findByPostId(postId).get();
             Cafe cafe = postCafe.getCafe();
-            MemberCafeTier memberCafeTier = memberCafeTierRepository.findByMemberIdAndCafeId(post.getMember().getId(),cafe.getId()).get();
-            detailResponseDto.updateDto(cafe.getName(), cafe.getBrandType(),memberCafeTier.getExp());
-            System.out.println("인증 빌드 완료");
+            MemberCafeTier memberCafeTier = memberCafeTierRepository.findByMemberIdAndCafeId(post.getMember().getId(), cafe.getId()).get();
+            detailResponseDto.updateDto(cafe.getName(), cafe.getBrandType(), memberCafeTier.getExp());
         }
         return detailResponseDto;
     }
 
     /**
-     * 7. 게시글 좋아요
+     * 5. 게시글 좋아요
      **/
     @Override
-    public PostLikeResponseDto likePost(PostLikeRequestDto likeRequestDto) throws Exception {
+    public PostLikeResponseDto likePost(PostLikeRequestDto likeRequestDto) {
 
         //1. 유저 확인
         CheckedResponseDto checked = postUtil.checkMember();
@@ -310,12 +313,14 @@ public class PostServiceImpl implements PostService {
         Boolean responseIsChecked;
 
         Optional<Post> postOptional = postRepository.findById(postId);
-        Post post;
-        post = postOptional.orElseThrow();
-
-        System.out.println("isChecked : " + likeRequestDto.getIsChecked());
+        if (postOptional.isEmpty() || postOptional == null) {
+            throw new PostException(PostExceptionType.BAD_POST_ID);
+        }
+        Post post = postOptional.get();
+        Optional<PostLike> postLikeOptional = postLikeRepository.findByPostIdAndMemberId(postId, memberId);
         if (!isChecked) {// 눌려있지 않다면 누른다음에, 좋아요 DB에 추가를 한다
-            if (postLikeRepository.findByPostIdAndMemberId(postId, memberId) == null) { // null 이면 맞음
+
+            if (postLikeOptional.isEmpty() || postLikeOptional == null) { // null 이면 맞음
                 PostLike postLike = PostLike.PostLikeBuilder()
                         .post(post)
                         .memberId(memberId)
@@ -323,14 +328,15 @@ public class PostServiceImpl implements PostService {
                 postLikeRepository.save(postLike);
                 // false -> true 로 바꿔서 반환
                 responseIsChecked = true;
-                System.out.println("postLike 생성완료");
             } else {// 유효성체크 : 안눌러져있는데, 또 누르기요청 : 에러발생
-                System.out.println("postLike 에러");
-                return null;
+                throw new PostException(PostExceptionType.POST_LIKE_CHECK_FAIL);
             }
         } else {
             // true - > DB 의 값을 삭제한다
-            System.out.println("postLike 삭제");
+
+            if (postLikeOptional == null || postLikeOptional.isEmpty()) {
+                throw new PostException(PostExceptionType.POST_LIKE_CHECK_FAIL);
+            }
             postLikeRepository.deleteByPostIdAndMemberId(postId, memberId);
             // true -> false
             responseIsChecked = false;
@@ -345,14 +351,21 @@ public class PostServiceImpl implements PostService {
         return response;
     }
 
+    /**
+     * 6. 게시글 업데이트로 가기
+     **/
     @Override
     public PostUpdateResponseDto updatePost(Long postId) {
-        Post post;
-        post = postRepository.findById(postId).get();
+
+        Optional<Post> postOptional = postRepository.findById(postId);
+        if (postOptional == null || postOptional.isEmpty()) {
+            throw new PostException(PostExceptionType.BAD_POST_ID);
+        }
+        Post post = postOptional.get();
         List<PostImage> postImageList = post.getPostImageList();
-        List<Map.Entry<String, String>> imgList = new ArrayList<>();
+        List<Map.Entry<Long, String>> imgList = new ArrayList<>();
         for (PostImage postImage : postImageList) {
-            imgList.add(new AbstractMap.SimpleEntry<>(postImage.getAccessKey(), postImage.getImgUrl()));
+            imgList.add(new AbstractMap.SimpleEntry<>(postImage.getId(), postImage.getImgUrl()));
         }
         PostUpdateResponseDto postUpdateResponseDto = PostUpdateResponseDto.builder()
                 .postId(postId)
@@ -362,29 +375,37 @@ public class PostServiceImpl implements PostService {
                 .build();
 
         return postUpdateResponseDto;
-
     }
 
     /**
-     * 2. 글 업데이트 [ 테스트 완료 ]
+     * 7. 글 업데이트 하기 [ 테스트 완료 ]
      **/
     @Override
-    public boolean updatePostForm(MultipartFile[] files, PostUpdateFormRequestDto updateDto) throws Exception {
+    public boolean updatePostForm(MultipartFile[] files, PostUpdateFormRequestDto updateDto) {
 
         // 1. 글 업데이트
         Long postId = updateDto.getPostId();
         String content = updateDto.getContent();
-        List<String> keyNameList = updateDto.getKeyNameList();
+        List<Long> imageIdList = updateDto.getImageIdList();
+
+        if (imageIdList != null || content != null || files != null) {
+
+        } else {
+            throw new PostException(NO_CONTENT_POST_FORM);
+        }
 
         Optional<Post> updateResult = postRepository.findById(postId);
         if (updateResult == null || updateResult.isEmpty()) {
-            return false;
+            throw new PostException(PostExceptionType.BAD_POST_ID);
         }
-        Post post = updateResult.orElseThrow();
-        post.updateContents(content);
+        Post post = updateResult.get();
+        if (content != null || !content.isEmpty()) {
+            post.updateContent(content);
+
+        }
 
         // 2. 이미지 업데이트
-        postUtil.imageDelete(post, keyNameList);
+        postUtil.imageDelete(post, imageIdList);
         if (files != null) {
             List<PostImage> postImages = postUtil.imageUpload(files);
             for (PostImage postImage : postImages) {
